simd -> binary search
Reduced dependencies and (seemingly) higher performance. I think the
arrays here are so short that SIMD may not even make sense? Large
improvements (5%-80%) across the board, especially when the length is >
capacity/2. Slowdowns (~20%) for `u8x24` at smaller lengths (not sure
why that type specifically).

Breaking my own rule of including criterion results since they _all_
changed, and it'd be very verbose.

Keeps the `simd` feature as an alias for `specialization` so we can
continue on `0.1`.

Closes #1 since no more SIMD.